@@ -1,19 +1,33 @@
+[tool.poetry]
+name = "python-code-jam-2024"
+version = "0.1.0"
+description = ""
+authors = ["Steve Mostovoy <stevemostovoysm@gmail.com>"]
+readme = "README.md"
+package-mode = false
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "3.12.*"
+pygame = "2.6.0"
+pillow = "10.4.0"
+setuptools = "^71.0.4"
+
+[tool.poetry.dev-dependencies]
+ruff = "0.5.4"
+pre-commit = "3.7.1"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
 [tool.ruff]
-# Increase the line length. This breaks PEP8 but it is way easier to work with.
-# The original reason for this limit was a standard vim terminal is only 79 characters,
-# but this doesn't really apply anymore.
 line-length = 119
-# Target Python 3.12. If you decide to use a different version of Python
-# you will need to update this value.
 target-version = "py312"
-# Automatically fix auto-fixable issues.
 fix = true
-# The directory containing the source code. If you choose a different project layout
-# you will need to update this value.
 src = ["src"]
 
 [tool.ruff.lint]
-# Enable all linting rules.
 select = ["ALL"]
 # Ignore some of the most obnoxious linting errors.
 ignore = [
@@ -29,6 +43,8 @@ ignore = [
     # Docstring punctuation.
     "D415",
     # Docstring quotes.
+    "D101",
+    "D103",
     "D301",
     # Builtins.
     "A",
@@ -39,26 +55,41 @@ ignore = [
     "TD003",
     "FIX",
     # Annotations.
+    "ANN001",
     "ANN101",
     "ANN102",
+    "ANN201",
+    "ANN202",
+    "ANN204",
+    "ANN205",
+    # Random
+    "S311",
+    # Docstrings
+    "D102",
+    # Magic values
+    "PLR2004",
+    "FBT002",
+    "FBT003",
+    # Unused argument
+    "ARG002",
+    # Complexity
+    "C901",
+    "PLR0912",
+    "PLR0913",
+    "PLR0915",
+    # Ambiguous formatting rules
+    "W191",
+    "E111",
+    "E114",
+    "E117",
+    "D206",
+    "D300",
+    "Q000",
+    "Q001",
+    "Q002",
+    "Q003",
+    "COM812",
+    "COM819",
+    "ISC001",
+    "ISC002",
 ]
-
-[tool.poetry]
-name = "python-code-jam-2024"
-version = "0.1.0"
-description = ""
-authors = ["Steve Mostovoy <stevemostovoysm@gmail.com>"]
-readme = "README.md"
-packages = [{include = "python_code_jam_2024"}]
-
-[tool.poetry.dependencies]
-python = "^3.10"
-pygame = "^2.6.0"
-pillow = "^10.4.0"
-discord.py = "^2.4.0"
-python-dotenv = "^1.0.1"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
