@@ -1,1 +1,3 @@
-from .ecosystem import *
+from .ecosystem import run_gif_generator, run_pygame
+
+__all__ = ["run_gif_generator", "run_pygame"]
